+++ R/focal.r
@@ -27,7 +27,7 @@
 #' 
 #' The center cell value is always included in the calculations, and all calculations ignore `NA` cells (i.e., they do not count as cells in the focal neighborhood).
 #'
-#' @param circle Logical: If `FALSE` (default), use a "square" neighborhood. If `TRUE`, use a "circular" neighborhood. When this is `TRUE`, argument `w` cannot be a matrix.
+#' @param circle Logical: If `FALSE` (default), use a square neighborhood. If `TRUE`, use a circular neighborhood. When this is `TRUE`, argument `w` cannot be a matrix.
 #'
 #' @param quantile Numeric between 0 and 1, inclusive: Quantile to calculate when `fun = "quantile"`. The default value is 0.5 (median), and valid values must be in the range between 0 and 1, inclusive.
 #'
@@ -83,8 +83,7 @@ methods::setMethod(
         size = size,
         nprocs = faster("cores"),
         memory = faster("memory"),
-        flags = c(.quiet(), "overwrite"),
-        intern = TRUE
+        flags = c(.quiet(), "overwrite")
     )
 
     if (circle) args$flags <- c(args$flags, "c")