+++ R/geomorphons.r
@@ -1,18 +1,16 @@
 #' Identify terrain feature types
 #'
 #' @description Geomorphons are idealized terrain types calculated from an elevator raster based on a moving window of a given size. The window is a torus (which can have an inner radius of 0, so can also be a circle), which allows it to identify geomorphons of a given size while ignoring ones larger or smaller. There are 10 basic geomorphons. Consult the The **GRASS** module [`r.geomorphon`](https://grass.osgeo.org/grass84/manuals/r.geomorphon.html) for more details and diagrams of each type of geomorphon. They include:
-#' 1 Flat areas: Focal area has approximately the same elevation as surrounding areas
-#' 2 Pits: An area is lower than all other surrounding areas
-#' 3 Valley: Focal area has elevation similar to two opposing side of the window but lower than the other two opposing sides
-#' 4 Footslope: Focal region is at the "bottom" of a slope
-#' 5 Hollow: A small valley/indention in the crest of a hill
-#' 6 Slope: Cells in the window form an approximately uniform slope
-#' 7 Spur: An extrusion at the foot of a hill (i.e.,, a small hill extending out from the foot of a slope)
-#' 8 Shoulder: The crest of a slope
-#' 9 Ridge: Opposite of a valley; focal area is higher than two opposing sides but approximately the same elevation as the other two opposing sides
-#' *10 Peak: Focal area is higher than any other in the window
-#'
-#' This function uses the "anglev2_distance" implementation of the `r.geomorphon` module.
+#' 1. Flat areas: Focal area has approximately the same elevation as surrounding areas
+#' 2. Pits: An area is lower than all other surrounding areas
+#' 3. Valley: Focal area has elevation similar to two opposing side of the window but lower than the other two opposing sides
+#' 4. Footslope: Focal region is at the "bottom" of a slope
+#' 5. Hollow: A small valley/indention in the crest of a hill
+#' 6. Slope: Cells in the window form an approximately uniform slope
+#' 7. Spur: An extrusion at the foot of a hill (i.e.,, a small hill extending out from the foot of a slope)
+#' 8. Shoulder: The crest of a slope
+#' 9. Ridge: Opposite of a valley; focal area is higher than two opposing sides but approximately the same elevation as the other two opposing sides
+#' 10. Peak: Focal area is higher than any other in the window
 #'
 #' @param x A single-layer `GRaster`, typically representing elevation.
 #'
@@ -25,7 +23,7 @@
 #' @param flatDist Numeric: Distance (in meters) to correct for the effect of large distances on the diminished capacity to identify "flat" geomorphons. If the distance between the focal area and a surrounding area surpasses this distance, then the effective value of `flat` will be reduced
 #'
 #' @param mode Character: Method for implementing the zenith/line-of-site search. Partial matching is used:
-#' * `"1"` (default): The "original" geomorphon mode (in **GRASS** module `r.geomorphon`, the "anglev1" method)
+#' * `"1"` (default): The "original" geomorphon mode (in **GRASS** module [`r.geomorphon`](https://grass.osgeo.org/grass84/manuals/r.geomorphon.html), the "anglev1" method)
 #' * `"2"`: Better handling of cases with equal zenith/nadir angles (the "anglev2" method)
 #' * `"2d"`: As `"2"`, but takes into account zenith/nadir distance ("anglev2_distance" method)
 #'