+++ man/examples/ex_streams.r
@@ -1,0 +1,16 @@
+if (grassStarted()) {
+
+# Setup
+library(terra)
+
+# Example data
+madElev <- fastData("madElev")
+
+# Convert a SpatRaster to a GRaster
+elev <- fast(madElev)
+
+# Calculate stream channels
+streams <- streams(elev)
+plot(streams)
+
+}