+++ NEWS.md
@@ -1,3 +1,20 @@
+# fasterRaster 8.3.0.7016 (2024-05-27)
+
+## Functionality
+o Added `streams()` for calculating location of stream channels from a DEM.
+o Added `terrainRuggednessIndex()` for calculating the terrain ruggedness index.
+o `unscale()` can skip unscaling of rasters by supplying `NA` in the `center` and/or `scale` vectors.
+o `writeRaster()` will now automatically choose the "least-lossy" `datatype` for a stack of rasters.
+o More robust checking of whether a vector is topologically valid or not when using `fast()`, and added option to aggregate or disaggregate polygons to overcome the issue.
+
+## Bug fixes
+o `crop()` correctly sets westernmost coordinate (was inappropriately too far west, in some cases).
+o `extend()` works when the "extension" factor is a integer.
+o `GRaster`s can now be multiplied by, divide by, added to, or subtracted from `numeric`s in scientific notation format.
+o `hist()` now works with `factor` `GRaster`s.
+o `plot()` relies on `writeRaster()` for `datatype` (which is better).
+o `writeRaster()` saves all-`NA` rows and columns.
+
 # fasterRaster 8.3.0.7015 (2024-05-21)
 
 ## Bug fixes